Add economic impact of custody accounts
On each validator exposure, record the sum of custody accounts stake that is backing the validator.
Still, custody accounts are not counted in exposure.others, so cannot claim payouts.
When computing the payout for a given era, use the sum of exposure.total and exposure.custody as the denominator.
This means that the part of the payout that is allocated to custody accounts is taken into account, but it is
never paid out.